pyproject.toml
[build-system]
requires = ["setuptools>=65.0"]
build-backend = "setuptools.build_meta"

[project]
name = "hf-platform"
version = "0.1.0"
description = "Production-ready hedge fund platform backbone (workers, monitoring, risk, sentiment)"
authors = [
  { name="Your Name", email="your@email.com" }
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  # Core
  "redis>=5.0.0",
  "pandas>=2.2.0",
  "numpy>=1.26.0",
  "pyyaml>=6.0.1",

  # Risk / research
  "scipy>=1.13.0",
  "scikit-learn>=1.5.0",
  "networkx>=3.2.0",

  # Monitoring
  "opentelemetry-sdk>=1.25.0",
  "opentelemetry-exporter-otlp>=1.25.0",
  "prometheus-client>=0.20.0",

  # NLP (for sentiment worker)
  "nltk>=3.9",
  "spacy>=3.7.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-cov>=5.0",
  "ruff>=0.5.0",
  "mypy>=1.10",
  "black>=24.0.0"
]

[tool.ruff]
line-length = 100
select = ["E", "F", "I"]
ignore = []

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=platform --cov=workers --cov=research --maxfail=2 -q"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.mypy]
python_version = "3.11"
check_untyped_defs = true
disallow_untyped_defs = true
ignore_missing_imports = true
strict_optional = true